fix(store): adjust mock store to handle selectors with props
