+++ modules/store/testing/spec/BUILD
@@ -2,18 +2,13 @@ load("//tools:defaults.bzl", "jasmine_node_test", "ts_test_library")
 
 ts_test_library(
     name = "test_lib",
-    srcs = glob(
-        [
-            "*.ts",
-            "meta-reducers/**/*.ts",
-        ],
+    srcs = glob(["**/*.ts"],
     ),
     deps = [
         "//modules/store",
         "//modules/store/testing",
         "//modules/store/spec/fixtures",
         "@npm//rxjs",
-        "@npm//ts-snippet",
     ],
 )
